added validators and exceptions to Service layer

+++ service/src/main/java/com/epam/esm/service/impl/CertificateServiceImpl.java
@@ -5,16 +5,18 @@ import com.epam.esm.dto.converter.CertificateConverterDTO;
 import com.epam.esm.dto.query.CertificatePageQueryDTO;
 import com.epam.esm.entity.Certificate;
 import com.epam.esm.entity.Tag;
+import com.epam.esm.exception.ServiceException;
 import com.epam.esm.repository.CertificateRepository;
 import com.epam.esm.repository.TagRepository;
 import com.epam.esm.service.CertificateService;
 import com.epam.esm.service.util.CertificateParamsRequestUtil;
 import com.epam.esm.specification.CertificateAllSpecification;
 import com.epam.esm.specification.SqlSpecification;
+import com.epam.esm.validator.CertificateDTOValidator;
+import com.epam.esm.validator.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 import java.util.Set;
 
@@ -22,12 +24,14 @@ import java.util.Set;
 public class CertificateServiceImpl implements CertificateService {
     private final CertificateRepository certificateRepository;
     private final TagRepository tagRepository;
+    private final Validator<CertificateDTO> validator;
 
 
     @Autowired
-    public CertificateServiceImpl(CertificateRepository certificateRepository, TagRepository tagRepository) {
+    public CertificateServiceImpl(CertificateRepository certificateRepository, TagRepository tagRepository, CertificateDTOValidator validator) {
         this.certificateRepository = certificateRepository;
         this.tagRepository = tagRepository;
+        this.validator = validator;
     }
 
     @Override
@@ -40,10 +44,19 @@ public class CertificateServiceImpl implements CertificateService {
     @Override
     @Transactional
     public CertificateDTO create(CertificateDTO certificateDTO) {
+        validator.cleanErrorMessages();
+        if (!validator.isValid(certificateDTO)) {
+            throw new ServiceException(validator.getErrorMessage());
+        }
+        String certificateName = certificateDTO.getName();
+        if (!certificateRepository.getByName(certificateName).isPresent()) {
+            throw new ServiceException("The certificate with this name (" + certificateName + ") is already exists.");
+        }
         Certificate certificate = CertificateConverterDTO.convertToEntity(certificateDTO);
         checkTags(certificate);
         certificate = certificateRepository.save(certificate);
         tagRepository.createCertificateTags(certificate);
+        certificate = certificateRepository.getById(certificate.getId());
         addTagsToCertificate(certificate);
         return CertificateConverterDTO.convertToDto(certificate);
     }
@@ -51,15 +64,23 @@ public class CertificateServiceImpl implements CertificateService {
     @Override
     @Transactional
     public CertificateDTO update(CertificateDTO certificateDTO) {
+        validator.cleanErrorMessages();
+        if (!validator.isValid(certificateDTO)){
+            throw new ServiceException(validator.getErrorMessage());
+        }
+        String certificateName = certificateDTO.getName();
+        if (!certificateRepository.getByName(certificateName).isPresent()) {
+            throw new ServiceException("The certificate with this name (" + certificateName + ") is already exists.");
+        }
         Certificate certificate = CertificateConverterDTO.convertToEntity(certificateDTO);
         certificateRepository.deleteCertificateTags(certificate.getId());
         checkTags(certificate);
         certificate = certificateRepository.update(certificate);
         tagRepository.createCertificateTags(certificate);
+        certificate = certificateRepository.getById(certificate.getId());
         addTagsToCertificate(certificate);
         return CertificateConverterDTO.convertToDto(certificate);
     }
-    //todo: find from db
 
     private void checkTags(Certificate certificate) {
         Set<Tag> tags = certificate.getTags();
@@ -75,7 +96,6 @@ public class CertificateServiceImpl implements CertificateService {
 
     @Override
     public List<CertificateDTO> executeQueryDTO(CertificatePageQueryDTO queryDTO) {
-
         CertificateParamsRequestUtil paramsRequestUtil = new CertificateParamsRequestUtil(queryDTO);
         String whereSQL = paramsRequestUtil.getWhereQueryWithParams();
         String sortSQL = paramsRequestUtil.getSortQueryWithParams();
